add new interface

+++ src/Nncase.Core/IR/NN/AttentionKVCache.cs
@@ -10,7 +10,7 @@ using System.Threading.Tasks;
 
 namespace Nncase.IR.NN;
 
-public enum AttentionCacheKind : byte
+public enum AttentionCacheKind : int
 {
     Key,
     Value,
